viewer finally works with multiple sets
